requirements.txt
attrs==21.2.0
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
protobuf==3.17.3
python-dotenv==0.19.0
six==1.16.0
Werkzeug==2.0.1
